requirements.txt
streamlit
tensorflow
scikit-learn
numpy
pandas
matplotlib
seaborn
joblib
requests
python-dotenv